+++ src/nunit.analyzers.sln
@@ -32,6 +32,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docs", "docs", "{2F501E14-3
 		..\documentation\NUnit1018.md = ..\documentation\NUnit1018.md
 		..\documentation\NUnit1019.md = ..\documentation\NUnit1019.md
 		..\documentation\NUnit1020.md = ..\documentation\NUnit1020.md
+		..\documentation\NUnit1021.md = ..\documentation\NUnit1021.md
+		..\documentation\NUnit1022.md = ..\documentation\NUnit1022.md
+		..\documentation\NUnit1023.md = ..\documentation\NUnit1023.md
+		..\documentation\NUnit1024.md = ..\documentation\NUnit1024.md
+		..\documentation\NUnit1025.md = ..\documentation\NUnit1025.md
 		..\documentation\NUnit2001.md = ..\documentation\NUnit2001.md
 		..\documentation\NUnit2002.md = ..\documentation\NUnit2002.md
 		..\documentation\NUnit2003.md = ..\documentation\NUnit2003.md
