+++ src/nunit.analyzers/TestCaseSourceUsage/TestCaseSourceUsesStringAnalyzer.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
 using Microsoft.CodeAnalysis;
@@ -152,32 +151,13 @@ namespace NUnit.Analyzers.TestCaseSourceUsage
             }
             else
             {
-                var sourceIsAccessible = context.SemanticModel.IsAccessible(
-                    syntaxNode.GetLocation().SourceSpan.Start,
-                    symbol);
-
-                if (attributeInfo.IsStringLiteral && sourceIsAccessible)
-                {
-                    var nameOfClassTarget = attributeInfo.SourceType.ToMinimalDisplayString(
-                        context.SemanticModel,
-                        syntaxNode.GetLocation().SourceSpan.Start);
-
-                    var nameOfTarget = attributeInfo.SourceType == context.ContainingSymbol.ContainingType
-                        ? stringConstant
-                        : $"{nameOfClassTarget}.{stringConstant}";
-
-                    var properties = new Dictionary<string, string>
-                    {
-                        { SourceCommonConstants.PropertyKeyNameOfTarget, nameOfTarget }
-                    };
-
-                    context.ReportDiagnostic(Diagnostic.Create(
-                        considerNameOfDescriptor,
-                        syntaxNode.GetLocation(),
-                        properties.ToImmutableDictionary(),
-                        nameOfTarget,
-                        stringConstant));
-                }
+                SourceHelpers.ReportToUseNameOfIfApplicable(
+                    context,
+                    syntaxNode,
+                    attributeInfo,
+                    symbol,
+                    stringConstant,
+                    considerNameOfDescriptor);
 
                 if (!symbol.IsStatic)
                 {
